requirements.txt
Flask
python-dotenv
groq
PyPDF2
Flask-Cors